added PostsPage compoenent
