+++ instagram/package.json
@@ -3,8 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "fuse.js": "^3.4.2",
     "react": "^16.8.2",
     "react-dom": "^16.8.2",
+    "react-fuse": "^0.3.7",
     "react-scripts": "2.1.5"
   },
   "scripts": {
