--- instagram/src/App.css
@@ -1,4 +1,0 @@
-h1 {
-   width:40%;
-	margin: 20px auto;
-}
