+++ instagram/src/components/PostsPage/PostsPage.css
@@ -1,0 +1,4 @@
+h1 {
+   width:40%;
+	margin: 20px auto;
+}
